Create api module in agency-client
