+++ agency_client/src/messages/a2a_message.rs
@@ -1,4 +1,4 @@
-use serde::{de, Deserialize, Deserializer, ser, Serialize, Serializer};
+use serde::{de, Deserialize, Deserializer};
 use serde_json::Value;
 
 use crate::message_type::{MessageFamilies, MessageType};
