+++ agency_client/src/messages/update_com_method.rs
@@ -1,7 +1,7 @@
 use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 use serde_json::Value;
-use serde::de::Error;
-use crate::agent_utils::ComMethod;
+
+use crate::api::agent::ComMethod;
 use crate::message_type::MessageType;
 use crate::messages::a2a_message::A2AMessageKinds;
