+++ aries_vcx/src/utils/provision.rs
@@ -1,4 +1,4 @@
-use agency_client::agent_utils;
+use agency_client::api::onboarding;
 
 use crate::error::prelude::*;
 use crate::libindy::utils::signus;
@@ -34,7 +34,7 @@ pub async fn provision_cloud_agent(provision_agent_config: &AgentProvisionConfig
     settings::get_agency_client_mut().unwrap().set_my_pwdid(&my_did);
     settings::get_agency_client_mut().unwrap().set_agent_vk(&provision_agent_config.agency_verkey); // This is reset when connection is established and agent did needs not be set before onboarding
 
-    let (agent_did, agent_vk) = agent_utils::onboarding(&my_did, &my_vk, &provision_agent_config.agency_did).await?;
+    let (agent_did, agent_vk) = onboarding::onboarding(&my_did, &my_vk, &provision_agent_config.agency_did).await?;
 
     Ok(AgencyClientConfig {
         agency_did: provision_agent_config.agency_did.clone(),
